Finops_ChatBot/werize_poc/requirements.txt
anthropic
gradio
fastapi
uvicorn
pinecone
sentence-transformers
python-dotenv
requests
